+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * #34: Fixed `?w` for PgSQL (argument of WHERE must be type boolean)
 * #31, #32: Fix error messages for PgSQL
+* Table: accumulate INSERT
 
 ## 2.2.1 (03.07.2014)